(category all) connect backend to frontend
all category (main, sub, child) connect to frontend menu

+++ app/Http/Controllers/Backend/SubCategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Backend;
 use App\DataTables\SubCategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\ChildCategory;
 use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use Str;
@@ -101,9 +102,13 @@ class SubCategoryController extends Controller
     public function destroy(string $id)
     {
         $subCategory = subCategory::findOrFail($id);
+        $childCategory = ChildCategory::where('sub_category_id', $subCategory->id)->count();
+        if($childCategory > 0) {
+            return response(['status' => 'error', 'message' =>'Ta kategoria ma pod sobą pod pod kategorie']);
+        }
         $subCategory->delete();
 
-        return response(['status' => 'success', 'message' =>'Usunięto pod kategorie']);
+        return response(['status' => 'success', 'Usunięto kategorie']);
     }
     public function changeStatus(Request $request)
     {
